modified the speex quality variable to be an int instead of uint8_t
git-svn-id: http://svn.mangler.org/mangler/trunk@330 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ libventrilo3/libventrilo3.c
@@ -632,7 +632,7 @@ _v3_recv(int block) {/*{{{*/
                                         uint8_t sample[codec->samplesize];
                                         SpeexBits bits;
                                         int encoded_size;
-                                        uint8_t tmp;
+                                        int tmp;
                                         static int rate = -1;
                                         static int format = -1;
                                         _v3_msg_0x52_speexdata *speexdata = malloc(sizeof(_v3_msg_0x52_speexdata));
